Points in the data set all implemented an interface from ScatterPlot. Break the dependency, there should be no relationship between DataSet points and ScatterPlot points. Change: 139082849
